LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-gainnode-interface/idl-test-expected.txt

CONSOLE MESSAGE: line 482: callback not yet supported
CONSOLE MESSAGE: line 482: callback not yet supported
dictionary GainOptions : AudioNodeOptions {
             float gain = 1.0;
};

[Constructor(BaseAudioContext context, optional GainOptions options)]
interface GainNode : AudioNode {

    readonly attribute AudioParam gain;

};

PASS webaudio Gain interfaces 
PASS GainNode interface: existence and properties of interface object 
FAIL GainNode interface object length assert_equals: wrong value for GainNode.length expected 1 but got 0
PASS GainNode interface object name 
PASS GainNode interface: existence and properties of interface prototype object 
PASS GainNode interface: existence and properties of interface prototype object's "constructor" property 
PASS GainNode interface: existence and properties of interface prototype object's @@unscopables property 
PASS GainNode interface: attribute gain 
PASS GainNode must be primary interface of gain_node 
PASS Stringification of gain_node 
PASS GainNode interface: gain_node must inherit property "gain" with the proper type 
